Fix JSON payload handling for requests with payload
The Heroku API client actually handles encoding of the body for us, so
when we were passing the unparsed input string directly, that was
being sent as a JSON *string* containing the full JSON payload. That
is, instead of sending the bytes

  { "foo": "bar" }

we were sending the bytes

  "{\"foo\": \"bar\" }"

and while technically this is valid JSON, it is not what the user
intended nor what the endpoint expects.

Instead, we parse the input and pass the resulting object to the
request, letting the client handle re-encoding for us.